Refactor video encoding integration across renderers
- Introduced a new `video_encoding.py` module containing `VideoEncodingConfig` and `_FFmpegVideoWriter` classes for unified video encoding configuration and writing.
- Updated `Open3DRenderer` and `BaseOpenCVRenderer` to utilize the new video encoding utilities, enhancing video output customization and compatibility.
- Removed redundant video encoding classes from `open3d_renderer.py` and `opencv_base.py` to streamline the codebase.
